+++ lib/system/strs_v2.nim
@@ -201,6 +201,16 @@ proc prepareMutation*(s: var string) {.inline.} =
     let s = unsafeAddr s
     nimPrepareStrMutationV2(cast[ptr NimStringV2](s)[])
 
+proc nimAddStrV1(s: var NimStringV2; src: NimStringV2) {.compilerRtl, inl.} =
+  #if (s.p == nil) or (s.len+1 > s.p.cap and not strlitFlag):
+  prepareAdd(s, src.len)
+  appendString s, src
+
+proc nimDestroyStrV1(s: NimStringV2) {.compilerRtl, inl.} =
+  frees(s)
+
+proc nimStrAtLe(s: string; idx: int; ch: char): bool {.compilerRtl, inl.} =
+  result = idx < s.len and s[idx] <= ch
 
 func capacity*(self: string): int {.inline.} =
   ## Returns the current capacity of the string.
@@ -211,4 +221,4 @@ func capacity*(self: string): int {.inline.} =
     assert str.capacity == 42
 
   let str = cast[ptr NimStringV2](unsafeAddr self)
-  result = if str.p != nil: str.p.cap else: 0
+  result = if str.p != nil: str.p.cap and not strlitFlag else: 0